refactor predownload_model fixture
